LPS-60850 Remove any reference to translator and move it to a processor package
