+++ modules/apps/collaboration/wiki/wiki-api/src/main/java/com/liferay/wiki/processor/WikiPageRenameProcessor.java
@@ -12,24 +12,13 @@
  * details.
  */
 
-package com.liferay.wiki.engine.mediawiki.translator;
-
-import com.liferay.wiki.translator.BaseWikiPageRenameProcessor;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Modified;
+package com.liferay.wiki.processor;
 
 /**
  * @author Roberto Díaz
- * @author Daniel Sanz
  */
-public class WikiPageRenameMediaWikiProcessor
-	extends BaseWikiPageRenameProcessor {
+public interface WikiPageRenameProcessor {
 
-	@Activate
-	@Modified
-	public void activate() {
-		regexps.put("\\[\\[Image:@old_title@/", "[[Image:@new_title@/");
-	}
+	public String translate(String content, String title, String newTitle);
 
 }
