+++ modules/apps/collaboration/wiki/wiki-engine-html/src/main/java/com/liferay/wiki/engine/html/processor/WikiPageRenameHTMLProcessor.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.wiki.engine.html.translator;
+package com.liferay.wiki.engine.html.processor;
 
-import com.liferay.wiki.translator.BaseWikiPageRenameProcessor;
-import com.liferay.wiki.translator.WikiPageRenameProcessor;
+import com.liferay.wiki.processor.BaseWikiPageRenameProcessor;
+import com.liferay.wiki.processor.WikiPageRenameProcessor;
 
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
